requirements.txt
kivy>=2.1.0
pyttsx3>=2.90
pdfminer.six>=20221105
Pillow>=9.0.0
pyjnius>=1.4.0
kivymd>=1.0.0
requests>=2.28.0
